==> 03_polyp_activity_exploration.Rmd <==
---
title: "Polyp Activity. Data Exploration and Statistical Analysis"
author: "Rhondda Jones, Natalia Andrade and Ira Cooke"
date: "29/08/2017"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
library(ordinal)
library(MASS)
library(reshape2)
library(tidyverse)
library(stringr)
```

```{r}
Polyp_Act<-read.csv("raw_data/Porites_polypact.csv")
Polyp_Act_long<-melt(Polyp_Act, id = c('TANK','TIMEPOINT','Treatment', 'SOFT','HARD','COLONY','SP','TREAT'))
polyp = Polyp_Act_long
polyp<- polyp %>% rename(Activity=value)
table(polyp$Activity)                       
polyp$TANK = factor(polyp$TANK)
Time<-as.integer(str_match(polyp$variable, pattern = "DAY([0-9]+)")[,2])
polyp<- cbind(polyp,Time)
## make polyp activity an ordered factor
polyp$Activity = factor(polyp$Activity, levels = c("C", "P","O"), ordered=TRUE)
## create time categories to generate descriptive plots
#polyp$TimeCat = cut(polyp$Time, breaks=c(0,9.9,19.9,34.9,60))
#polyp$TimeCat = cut(polyp$Time, breaks=c(0,10.1,13.1,16.1,19.1,22.1,25.1,28.1,31.1,34.1,37.1,40.1,43.1,46.1,49.1,52.1,55.1,58.1,60))
polyp$TimeCat = cut(polyp$Time, breaks=c(0,14.1,19.1,24.1,29.1,34.1,39.1,44.1,49.1,54.1,60)) 
polyp$TimeCat = factor(polyp$TimeCat, labels=c("<15","15-20","20-25","25-30","30-35","35-40","40-45","45-50","50-55","55-60"))

```

```{r}
## table and plot for no competition
polyp<- polyp %>% group_by(HARD)
polyp_Pd<- polyp %>% group_by(HARD) %>% filter(HARD=="Pd")
polyp_Pe<- polyp %>% group_by(HARD) %>% filter(HARD=="Pe")
polyp_Pf<- polyp %>% group_by(HARD) %>% filter(HARD=="Pf")

timeTableN = with(polyp[polyp$TREAT=="NO",],table(TimeCat, Activity))
#pdf(file="results/MosaicPlot_Control_Allsamples.pdf",height =8,width = 10)
mosaicplot(timeTableN, col=1:3,cex.axis = 1.05, font=14,
           main="Activity proportions without competition", xlab="Time (days)")
#dev.off()
```


```{r}
timeTableN_Pd = with(polyp_Pd[polyp_Pd$TREAT=="NO",],table(TimeCat, Activity))
#pdf(file="results/MosaicPlot_Control_Pd.pdf",height =8,width = 12)
mosaicplot(timeTableN_Pd, col=1:3,cex.axis = 1.1, 
           main="Activity proportions without competition_Pd", xlab="Time (days)")
#dev.off()
```


```{r}
timeTableN_Pe = with(polyp_Pe[polyp_Pe$TREAT=="NO",],table(TimeCat, Activity))
#pdf(file="results/MosaicPlot_Control_Pe.pdf",height =8,width = 12)
mosaicplot(timeTableN_Pe, col=1:3,cex.axis = 1.1, 
           main="Activity proportions without competition_Pe", xlab="Time (days)")
#dev.off()
```


```{r}
timeTableN_Pf = with(polyp_Pf[polyp_Pf$TREAT=="NO",],table(TimeCat, Activity))
#pdf(file="results/MosaicPlot_Control_Pf.pdf",height =8,width = 12)
mosaicplot(timeTableN_Pf, col=1:3,cex.axis = 1.1, 
           main="Activity proportions without competition_Pf", xlab="Time (days)")
#dev.off()
```


```{r}
## table and plot for competition
par(mfrow=c(1,3))

timeTableY = with(polyp[polyp$TREAT=="Yes",],table(TimeCat, Activity))
#pdf(file="results/MosaicPlot_Treat_Allsamples.pdf",height =8,width = 10)
mosaicplot(timeTableY, col=1:3,cex.axis = 1.1, 
           main="Activity proportions with competition", xlab="Time (days)")
#dev.off()
```


```{r}
timeTableY_Pd = with(polyp_Pd[polyp_Pd$TREAT=="Yes",],table(TimeCat, Activity))
#pdf(file="results/MosaicPlot_Treat_Pd.pdf",height =8,width = 12)
mosaicplot(timeTableY_Pd, col=1:3, cex.axis = 1.1, 
           main="Activity proportions with competition_Pd", xlab="Time (days)")
#dev.off()
```


```{r}
timeTableY_Pe = with(polyp_Pe[polyp_Pe$TREAT=="Yes",],table(TimeCat, Activity))
#pdf(file="results/MosaicPlot_Treat_Pe.pdf",height =8,width = 12)
mosaicplot(timeTableY_Pe, col=1:3, cex.axis = 1.1, 
           main="Activity proportions with competition_Pe", xlab="Time (days)")
#dev.off()
```


```{r}
timeTableY_Pf = with(polyp_Pf[polyp_Pf$TREAT=="Yes",],table(TimeCat, Activity))
#pdf(file="results/MosaicPlot_Treat_Pf.pdf",height =8,width = 12)
mosaicplot(timeTableY_Pf, col=1:3, cex.axis = 1.1, 
           main="Activity proportions with competition_Pf", xlab="Time (days)")
#dev.off()
```


```{r}
# check data balance
table(polyp$Activity,polyp$TANK)
table(polyp$TREAT,polyp$TANK)
```

```{r}
## random effect Tank
## potential explanatory variables: Time, TREAT, HARD

## initially try models without random effects
fm00 = clm(Activity ~ Time+TREAT+HARD, Hess=TRUE, data=polyp)
summary(fm00)
```


```{r}
slice.fm00 = ordinal::slice(fm00, lambda = 5)
par(mfrow=c(2,3))
plot(ordinal::slice(fm00))     ## well-behaved for both thresholds & explanatory variables
```


```{r}
## curvature?
fm01 = clm(Activity ~ Time+I((Time/10)^2) + TREAT +HARD, Hess=TRUE, data=polyp)
summary(fm01)
anova(fm00, fm01)
confint(fm01)
## clear curvature (changes flatten off as day increases)
```


```{r}
## well-behaved
##  better to use time categories rather than continuous time?
fm02 = clm(Activity ~ TimeCat+TREAT+HARD, Hess=TRUE, data=polyp)
summary(fm02)
anova(fm01, fm02)
## categories are much better
slice.fm02 = ordinal::slice(fm02, lambda = 5)
par(mfrow=c(2,3))
plot(ordinal::slice(fm02)) ## still well behaved
```


```{r}
#   nominal instead?
#fm02.nom = clm(Activity ~ 1, nominal= ~TimeCat + TREAT+HARD, data=polyp)
#summary(fm02.nom)
#anova(fm01, fm02.nom)    ### nominal is NOT better for fixed effects
                         ### so proportional odds assumption is not violated


### INCLUDE RANDOM EFFECT OF TANK
fm1.1 = clmm2(location=Activity ~ Time+TREAT+HARD, Hess=TRUE, data=polyp)
fm1 = clmm2(location=Activity ~ Time+TREAT+HARD, random=TANK, Hess=TRUE, data=polyp)

summary(fm1)
AIC(fm1, fm1.1)
anova(fm1, fm1.1)
## definitely needs to allow for tank effects!!!

## check for curvature in the time effect (add quadratic term)
fm2 =clmm2(location=Activity ~ Time+I((Time/10)^2)+TREAT+HARD, random=TANK,Hess=TRUE,  data=polyp)
summary(fm2)
## HAS CURVATURE
```


```{r}
## Revert to time categories as before
fm3a =clmm2(location=Activity ~ 1, random=TANK, Hess=TRUE,  data=polyp)
       ## convergence failure for random effects but marginal - 0.00002
fm3b =clmm2(location=Activity ~ TimeCat, random=TANK, Hess=TRUE,  data=polyp)
fm3c =clmm2(location=Activity ~ TimeCat+TREAT, random=TANK, Hess=TRUE,  data=polyp)
fm3d =clmm2(location=Activity ~ TimeCat+TREAT+HARD, random=TANK, Hess=TRUE,  data=polyp)

anova(fm3a, fm3b, fm3c, fm3d)
AIC(fm3a, fm3b, fm3c, fm3d)
  ## fm3c best, residual deviance OK, fit looks OK

summary(fm3d) ## Hessian OK
```


```{r}
## check back for TANK effects with the time category model
fm3f = clmm2(location=Activity ~ TimeCat+TREAT+HARD, Hess=TRUE,  data=polyp)
anova(fm3f, fm3d)
AIC(fm3f, fm3d)
## TANK effect remains hugely significant!!


## check confidence intervals & behaviour of tank effect
pr3d = profile(fm3d, range=c(0.3, 2), nsteps=30, trace=0)
confint(pr3d)
plot(pr3d)
## not bad, sd pretty well estimated despite the warnings
```


```{r}
## Find predicted probabilities for an "average" tank
predictions = cbind(polyp, pred=predict(fm3d, newdata=polyp))
with(na.omit(predictions), round(tapply(pred, list(Activity, TREAT,HARD, TimeCat), mean), 3))

## whereas actual proportions are:
## for no competition
round(timeTableN_Pd/rowSums(timeTableN_Pd), 3)
## for competition
round(timeTableY_Pd/rowSums(timeTableY_Pd), 3)
## pretty good match ...


```

```{r}
ci <- fm3d$ranef + qnorm(0.975) * sqrt(fm3d$condVar) %o% c(-1, 1)
ord.re <- order(fm3d$ranef)
ci <- ci[order(fm3d$ranef),]
plot(1:36, fm3d$ranef[ord.re], axes=FALSE, ylim=range(ci),
       xlab="Tank", ylab="Tank effect")
axis(1, at=1:36, labels = ord.re)
axis(2)
for(i in 1:36) segments(i, ci[i,1], i, ci[i, 2])
 abline(h = 0, lty=2)
exp(coef(fm3d)[5])
 
```
